Forced fallbacklayer_norm() output to be in FP32

+++ src/mltools/functions.py
@@ -397,7 +397,7 @@ def fallbacklayer_norm(
     if bias is not None:
         _x += bias.to(nform)
 
-    return _x
+    return _x.float()
 
 
 def poly2gelu(xin, nform="float16"):
